Make Simple example build with base-4.18.* (GHC 9.6)
`Eq1` now has a quantified `Eq` superclass as of `base-4.18.*`. A knock-on
consequence of this change is that the TH-derived `Eq1` instance in the
`Simple` example no longer compiles, as the corresponding `Eq` instance is
defined in a separate top-level group than the `Eq` instance. A similar problem
exists for the `Ord1`, `Read1`, and `Show1` instances.  In fact, this problem
is not unique to the `Simple` example—it will affect any instance that follows
the templates given in the Haddocks for `Bound` and `Bound.TH`, both of which
make use of TH!

I have adapted to these changes by defining all of the TH-derived instances
(e.g., the `Eq1` instances) in the same TH splice as the hand-written instances
(e.g., the `Eq` instances). This is a bit of a heavy-handed solution, but I
can't think of a better one given the restrictions that TH splices impose.